--- src/lib/rooms/utils/index.ts
@@ -1,1 +1,0 @@
-export * from './middleware'
