+++ src/modules/rooms/components/RoomTabs.tsx
@@ -2,7 +2,7 @@ import type { Room } from 'types/rooms'
 import clsx from 'clsx'
 import toast from 'react-hot-toast'
 import { Tab } from '@headlessui/react'
-import Button from 'lib/ui/components/Button'
+import Button from 'modules/ui/components/Button'
 import Members from './Members'
 import Tracks from './Tracks'
 import { useGeneratePlaylistMutation } from '../services/roomApi'
