+++ src/modules/rooms/components/Rooms.tsx
@@ -2,8 +2,8 @@ import clsx from 'clsx'
 import Link from 'next/link'
 import Image from 'next/image'
 import toast from 'react-hot-toast'
-import Spinner from 'lib/ui/components/Spinner'
-import { useGetUserRoomsQuery } from 'lib/rooms/services/roomApi'
+import Spinner from 'modules/ui/components/Spinner'
+import { useGetUserRoomsQuery } from 'modules/rooms/services/roomApi'
 
 const Rooms = () => {
   const { data, error, isLoading } = useGetUserRoomsQuery()
