+++ src/modules/rooms/utils/baseQuery.ts
