+++ src/modules/rooms/utils/index.ts
@@ -1,0 +1,1 @@
+export * from './baseQuery'
