+++ src/modules/ui/constants/navigation.ts
