+++ src/pages/404.tsx
@@ -1,5 +1,5 @@
 import Image from 'next/image'
-import LandingLayout from 'lib/ui/layouts/LandingLayout'
+import LandingLayout from 'modules/ui/layouts/LandingLayout'
 
 const Custom404 = () => {
   return (
