+++ src/pages/_app.tsx
@@ -1,7 +1,7 @@
 import { SessionProvider } from 'next-auth/react'
 import type { AppProps } from 'next/app'
 import { Provider } from 'react-redux'
-import store from 'lib/store'
+import store from 'modules/store'
 import 'styles/globals.css'
 
 function MyApp({ Component, pageProps }: AppProps) {
