+++ src/pages/api/auth/[...nextauth].ts
@@ -1,9 +1,9 @@
 import NextAuth from 'next-auth'
 import SpotifyProvider from 'next-auth/providers/spotify'
 import { MongoDBAdapter } from '@next-auth/mongodb-adapter'
-import clientPromise from 'core/mongodb'
-import { LOGIN_URL } from 'core/spotify'
-import { refreshAccessToken } from 'lib/server/utils'
+import clientPromise from 'lib/mongodb'
+import { LOGIN_URL } from 'lib/spotify'
+import { refreshAccessToken } from 'server/utils'
 
 export default NextAuth({
   adapter: MongoDBAdapter(clientPromise),
@@ -34,7 +34,7 @@ export default NextAuth({
       }
 
       if (Date.now() < token.accessTokenExpires) {
-        console.log('Existing Access token is valid')
+        // Existing Access token is valid
         return token
       }
