+++ src/server/utils/index.ts
@@ -1,4 +1,4 @@
 export * from './auth'
 export * from './authMiddleware'
+export * from './roomMiddleware'
 export * from './nc-options'
-export * from './baseQuery'
